Move Route Duration Annotations to NavigationMapView instead of Example project. Entails also moving a number of class extensions and image resources over as well.
