--- Extensions/UIImage+Example.swift
@@ -1,28 +1,0 @@
-import UIKit
-
-extension UIImage {
-    func tint(_ tintColor: UIColor) -> UIImage {
-        let imageSize = size
-        let imageScale = scale
-        let contextBounds = CGRect(origin: .zero, size: imageSize)
-
-        UIGraphicsBeginImageContextWithOptions(imageSize, false, imageScale)
-
-        defer { UIGraphicsEndImageContext() }
-
-        UIColor.black.setFill()
-        UIRectFill(contextBounds)
-        draw(at: .zero)
-
-        guard let imageOverBlack = UIGraphicsGetImageFromCurrentImageContext() else { return self }
-        tintColor.setFill()
-        UIRectFill(contextBounds)
-
-        imageOverBlack.draw(at: .zero, blendMode: .multiply, alpha: 1)
-        draw(at: .zero, blendMode: .destinationIn, alpha: 1)
-
-        guard let finalImage = UIGraphicsGetImageFromCurrentImageContext() else { return self }
-
-        return finalImage
-    }
-}
