+++ MapboxNavigation/Resources/Assets.xcassets/RouteAnnotations/Contents.json
