+++ MapboxNavigation/Resources/Assets.xcassets/RouteAnnotations/RouteInfoAnnotationLeftHanded.imageset/Contents.json
