+++ MapboxNavigation/Resources/Assets.xcassets/RouteAnnotations/RouteInfoAnnotationLeftHanded.imageset/LeftHandUpDown.pdf
