+++ MapboxNavigation/Resources/Assets.xcassets/RouteAnnotations/RouteInfoAnnotationRightHanded.imageset/Contents.json
