+++ MapboxNavigation/Resources/Assets.xcassets/RouteAnnotations/RouteInfoAnnotationRightHanded.imageset/RightHandUpDown.pdf
